Checked in Cpp1 passthrough tests
- And fixed a character literal recognition bug
- And increased line length (again) after encountering a(n even longer) very long line in real world code (I thought 19,518 chars was a long line until I came across 80,556...)

+++ common.h
@@ -59,7 +59,7 @@ struct source_line
 
 
 using lineno_t = int32_t;
-using colno_t  = int16_t;
+using colno_t  = int32_t;   // not int16_t... encountered >80,000 char line during testing
 
 struct source_position
 {
